requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
flask-limiter==3.5.0
flask-pymongo==2.3.0
pymongo==4.6.1
spacy==3.7.3
joblib==1.3.2
python-dotenv==1.0.1
Pillow==11.1.0
pytesseract==0.3.13
